scraper/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
extruct==0.14.0
html-text==0.5.2
html5lib==1.1
idna==3.4
isodate==0.6.1
jstyleson==0.0.2
lxml==4.9.1
mf2py==1.1.2
python-dotenv==0.21.0
pyparsing==3.0.9
pyRdfa3==3.5.3
rdflib==6.2.0
recipe-scrapers==14.23.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
w3lib==2.0.1
webencodings==0.5.1
